OpenSSL-Query: Implement late evaluation further in the REST client
In the base class, take note of all implementions, not just the first
that initiated successfully.  Instead, use all impllementations in the
methods, and return the value from the first successful one.

For the REST implementations, this means moving the check of server
errors to the implementation methods instead of doing so when building
the object.