Adapt gitaddrev to use OpenSSL::Query
IMPORTANT NOTE: with this change, it's important to install OpenSSL-Query.
If you have direct access to the databases, it's also important to install
the QueryApp libraries.  Finally, you must either defined the environment
variable BUREAU with the directory where the databases reside as value,
or defined the two variables CLADB and PERSONBD with the path of the
respective files as value.  If none of these variables are set, the files
are assumed to be in the current working directory.

For gitaddrev, the changes are substancial:

- Do CLA checks early
- Increase the CLA checks
- Add a check of number of reviewers
- Add a check of at least one OMC member
- List *all* unknown reviewers, not just the first
- With --list, show which ones are OMC members

For ghmerge, the setting of CLADB got removed.