gitaddrev: correct logic error for --myemail
If the --myemail value is someone with a CLA, they count as a reviewer
no matter what.  If they are a OMC member, they count as such no
matter what as well.  However, they still need to be removed from the
@reviewer list before the commit message is rewritten, since they
can't have a Reviewed-by line on their own commit.