gitaddrev: more logic corrections
Refactor reviewer addition to a function of its own.  It will return
the person's rev tag on success.

Furthermore, don't do any further checks on the --myemail argument,
just try to add the corresponding rev tag as well.

Finally, try to add the rev tag for the committer author, but
immediately register that it should be removed from @reviewers after
checking the reviewer count.  This ensures that the commit author
never will figure as reviewer as well, even if the addresses differ
from each other.