cherry-checker: initial commit
usage: cherry-checker [-h] [-a] [-s] [-r]

Shows the commits in 'master...OpenSSL_1_1_1-stable' which are eligible for
cherry-picking. A commit is considered cherry-picked, if there is another
commit on the "other side" which introduces an equivalent patch. For details,
see the documentation of the '--cherry-mark' option in the git-log(1) manpage.

optional arguments:
  -h, --help    show this help message and exit
  -a, --all     Show all commits, also those which have been cherry-picked.
  -s, --sort    Sort commits w.r.t. pull request number and author date.
  -r, --remote  Compare the remote branches instead of the local ones.

Reviewed-by: Richard Levitte <levitte@openssl.org>
(Merged from https://github.com/openssl/tools/pull/32)