release-tools/release.sh et al: introduce RELEASE_FILES and use it
This is a variable that holds a list of files other than the version file,
that need to be touched up with release information.  This variable is set
by the version functions, as the files are varied, depending on the versio
scheme and series.

We also add the fixup files for all possible release files.

Reviewed-by: Tomas Mraz <tomas@openssl.org>
(Merged from https://github.com/openssl/tools/pull/140)