Adapt documentation for the new release script, and clean up
release-tools/release.sh and associated files should now be a good enough
replacement of release-tools/mkrelease.pl and associated files.
Therefore, HOWTO-make-a-release.md is adapted to only refer to the new
script, and release-tools/mkrelease.pl and associated files are removed.

Someone might want to ask, why shell scripts rather than perl?
The reasoning is that the OpenSSL team does most if not all its development
on Unix-like systems, and the release script is essentially a wrapper around
diverse shell commands anyway, it therefore seems sensible to use the shell
language.