When checking for make update changes, only check tracked files
Sometimes, 'make update' leaves behind files that aren't tracked, and this
shouldn't cause trying to add an extra commit.  However, we only checked
with 'git status --porcelain', which does list untracked files.  This can
cause release failure, so we add a few more options so git status doesn't
report on these files that should be ignored.

This also includes a documenting comment for a piece of the scripts that's
potentially confusing without the explanation.