release-tools/publish-releases.sh: Change staging types to 'ordinary' and 'security'
Having the staging types 'public', 'premium' and 'security' was nonsense,
we don't really have that sort of division.  'ordinary' and 'security'
staging makes more sense for what we do, and that leaves it to the script
to fully determine for each version being released if it should become a
public or a premium release.