release-tools/{stage,publish}-releases.sh: Better gpg handling
gpg is very annoying with how it prompts for pass phrases.  However, it's
possible to tell it to use the terminal with '--pinentry-mode loopback'.
In turn, git is annoying with how one can't specify gpg options directly;
instead, a small script that does the right thing is produced, and passed
to git with the gpg.program config option.