Allow for setting of extra cflags
When I build to run these tests under perf, I need to include a few
extra cflags (notably -rpath an -fno-omit-frame-pointer).  They're not
generally needed, but it would be nice to be able to set them easily in
an environment.

Add CFLAGS and LDFLAGS support to allow for easy
configuration