Make use of new -s option to not share ssl_ctx objects
To better understand our performance behavior, add an option to the
handshake test to allow the threads to repeatedly allocate their own
contexts per connection.  If -s is specified this will be the behavior.
Omitting -s uses the default behavior, which is to share an ssl_ctx
among all connections