The hanshake and x509storeissuer tools also require a certsdir argument
We currently pass the certsdir argument as the first one, then
follows the threadcount argument. I'd like to change it so the
threadcount argument will be the last one. After changes in
this PR will land to repo, then all tools will accept a threadcount
as the last argument. This will simplify additional scripting
built on top of those tools.

This also unifies the getopt handling across the tools to make
it buildable on Windows.

Reviewed-by: Matt Caswell <matt@openssl.org>
Reviewed-by: Tomas Mraz <tomas@openssl.org>
(Merged from https://github.com/openssl/tools/pull/206)