Drop requirement for OTC review
Current bylaws under general policies:
https://github.com/openssl/general-policies/blob/master/policies/committer-policy.md

Indicate the following:
```
All submissions must be reviewed and approved by at least two committers.
Neither of the reviewers can be the author of the submission.
```

Given that, it seems we should adjust our tooling to reflect the lack of
need for an OTC approval